requirements.txt
fastapi
uvicorn
psycopg2
SQLAlchemy
asyncpg
pydantic
pyyaml
PyJWT
pytest-cov